+++ graph_layout/example/graph_layout_example.dart
@@ -1,6 +1,1 @@
-import 'package:graph_layout/graph_layout.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}
+// TODO: Add example.