+++ graph_layout/lib/graph_layout.dart
@@ -1,8 +1,4 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library graph_layout;
 
-export 'src/graph_layout_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'src/data_structures.dart';
+export 'src/spring.dart';