+++ graph_display/CHANGELOG.md
@@ -1,1 +1,3 @@
 ## Next
+
+* Add widget `SpringGraphDisplay` for displaying graphs using a spring based layout algorithm.