+++ graph_layout/CHANGELOG.md
@@ -1,1 +1,5 @@
 ## Next
+
+* Add basic data structures: `Node`, `IntegerNode`, `Edge` (undirected), `Graph`, `LayoutAlgorithm`.
+* Add basic typedefs: `AdjacencyList`, `EdgeList`, `NodeLayout`.
+* Add spring based layout algorithm `SpringSystem`, fully implemented in Dart.